feat: Migrated project to Typescript and eslint and prettier added
