+++ .eslintrc.js
@@ -1,0 +1,28 @@
+module.exports = {
+  plugins: ['prettier', 'react', '@typescript-eslint'],
+  extends: [
+    'plugin:@typescript-eslint/eslint-recommended',
+    'plugin:@typescript-eslint/recommended',
+    'prettier',
+  ],
+  rules: {
+    'prettier/prettier': [
+		  'error',
+		  {
+		    'endOfLine': 'lf',
+		  }
+		],
+    'react/jsx-filename-extension': [
+      1,
+      { extensions: ['.js', '.jsx', '.tsx'] },
+    ],
+  },
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: ['.js', '.jsx', '.ts', '.tsx'],
+        paths: ['src'],
+      },
+    },
+  },
+}
