+++ package.json
@@ -3,13 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@apollo/client": "^3.5.10",
     "@apollo/react-hooks": "^4.0.0",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
+    "@types/node": "^17.0.21",
+    "@types/react": "^17.0.40",
+    "@types/react-dom": "^17.0.13",
     "apollo-boost": "^0.4.9",
     "apollo-cache-inmemory": "^1.6.6",
-    "apollo-client": "^2.6.10",
     "apollo-link-ws": "^1.0.20",
     "graphql": "^16.3.0",
     "graphql-tag": "^2.12.6",
@@ -17,6 +20,7 @@
     "react-dom": "^17.0.2",
     "react-scripts": "5.0.0",
     "subscriptions-transport-ws": "^0.11.0",
+    "typescript": "^4.6.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -43,5 +47,12 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.14.0",
+    "@typescript-eslint/parser": "^5.14.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "kcd-scripts": "^12.1.0",
+    "prettier": "^2.5.1"
   }
 }
