+++ prettier.config.js
@@ -1,0 +1,11 @@
+// const prettierConfig = require('kcd-scripts/prettier');
+
+// module.exports = Object.assign(prettierConfig, {
+//   bracketSpacing: true,
+//   endOfLine: 'lf',
+//   tabWidth: 2,
+//   useTabs: true,
+//   semi: true,
+//   bracketSameLine: false,
+//   printWidth: 250,
+// });
