--- src/pages/TestingPage.js
@@ -1,3 +1,0 @@
-export const Testing = () => (
-	<h1>Hello Apollo Client</h1>
-);
