+++ src/pages/TestingPage.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const Testing = () => <h1>Hello Apollo Client</h1>;
